Update Two sum-II.java

+++ Blind 75/Two pointer/Two sum-II.java
@@ -11,7 +11,7 @@ class Solution {
         int high = numbers.length - 1;
         
         while(low< high){
-            int sum = numbers[low] + numbers[high];
+            long sum = numbers[low] + numbers[high];
             
             if(sum == target){
                 return new int[]{low+1 , high +1};
